use imzml file mz dtype to produce ds segments of required size (5MB default)
* minor modules restructure
